back to help-version
`meow` may be great for some, but if all I want is handling `--help`
and `--version` arguments and showing help with the given exit code,
then `meow` is a bit too complicated for that.

What I also don't like about it is that it bundles `minimist` for some
reason - that makes it too frameworkish and prevents me from using other
argument parsers or using minimist differently (there are even
`minimist-plugins` and `minimist-events` projects that I might want
to use in the future, but wouldn't be able to with `meow`).

After long search I finally arrived at `help-version`, back to where I
started. So here we are.